# backend1/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Contactos",
    "description": "Documentación de la API para gestionar contactos.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Obtener todos los contactos",
        "description": "Devuelve una lista de todos los contactos almacenados.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Lista de contactos",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Crear un nuevo contacto",
        "description": "Crea un nuevo contacto con los datos proporcionados.",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contacto creado exitosamente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Obtener un contacto por ID",
        "description": "Devuelve los detalles de un contacto específico.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del contacto"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del contacto",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "404": {
            "description": "Contacto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Actualizar un contacto por ID",
        "description": "Actualiza los datos de un contacto existente.",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del contacto"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contacto actualizado exitosamente"
          },
          "404": {
            "description": "Contacto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Eliminar un contacto por ID",
        "description": "Elimina un contacto existente por su ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del contacto"
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto eliminado exitosamente"
          },
          "404": {
            "description": "Contacto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {      
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-01-01"
        }
      },
      "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"]
    }
  }
}